+++ RepairSystem-Backend/src/main/java/ca/mcgill/ecse321/repairsystem/model/AdministrativeAssistant.java
@@ -7,7 +7,7 @@ import javax.persistence.Entity;
 public class AdministrativeAssistant extends Person
 {
 	
-	public AdministrativeAssistant(String aName, int id, String aPassword, int aPhone, String aEmail) {
+	public AdministrativeAssistant(String aName, int id, String aPassword, long aPhone, String aEmail) {
 		super(aName,id, aPassword, aPhone, aEmail);
 	}
